some refactoring and styling changes
